requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
aiohttp==3.11.12
pycryptodome==3.21.0